Improve logic for guessing date format
Use the fact that all the dates are chronological and past dates to try to make
a best guess for the date format in the exported data. We assume that the month
in the date list should change less frequently than the days, when multiple
formats seem like valid ones, based on the available dates.